fix(meeting-detection): decouple parsers from native binding in tests
The classifier logic lived alongside the native-watcher wiring, so testing
the pure parsers transitively required the napi binding (binding.cjs) at
import time. That artifact is gitignored and not built in CI, causing the
release check job to crash with 'Cannot find module binding.cjs'.

Extract the pure classifiers into macos-classify.ts/windows-classify.ts with
no native imports, keeping macos.ts/windows.ts as thin wrappers that wire them
to the native watcher. Move tests to the classifier modules and replace the
file-level process.platform guards with explicit test.skipIf, so off-platform
skips are visible in test output instead of silently dropping the suite.
